Added new shop drawables, and changed call api.

+++ app/src/main/java/com/gruppe17/madbudget/coop/model/CoopStore.kt
@@ -29,7 +29,7 @@ class CoopStore(
 
 class CoopLocation (
     val type: String, // Useless
-    val coordinates: List<Int>
+    val coordinates: List<Double>
 )
 
 class CoopOpeningHour(
